+++ packages/zapp-core/src/interfaces/IProcessor.ts
@@ -1,3 +1,3 @@
-export interface IProcessor {
-  (output: string): Promise<string>;
+export interface IProcessor<T = any> {
+  (options?: T): (output: string) => Promise<string>;
 }