+++ packages/zapp-prettier/src/processors/PrettierProcessor.ts
@@ -1,8 +1,16 @@
 import { IProcessor } from '@zappjs/core';
-import { format, resolveConfig } from 'prettier';
+import { Options, format } from 'prettier';
 
-export const PrettierProcessor: IProcessor = async (input: string) => {
-  const options = await resolveConfig('./.prettierrc.json');
-  const output = format(input, options ? { parser: 'babel-ts', ...options } : undefined);
-  return output;
+export const PrettierProcessor: IProcessor<Options> = (options) => {
+  return async (input: string) => {
+    const output = format(input, {
+      parser: 'babel-ts',
+      printWidth: 100,
+      singleQuote: true,
+      tabWidth: 2,
+      useTabs: false,
+      ...options,
+    });
+    return output;
+  };
 };