+++ examples/next-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # next-api
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @zappjs/next-zapp@0.1.3
+
 ## null
 
 ### Patch Changes