+++ packages/zapp-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zappjs/core
 
+## 2.1.0
+
+### Minor Changes
+
+- updates processor to accept options
+
 ## 2.0.0
 
 ### Minor Changes