+++ packages/zapp-git/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zappjs/git
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zappjs/core@2.1.0
+  - @zappjs/handlebars@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes