+++ packages/zapp-handlebars/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zappjs/handlebars
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zappjs/core@2.1.0
+
 ## 0.2.1
 
 ### Patch Changes