+++ packages/zapp-typescript/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zappjs/typescript
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @zappjs/core@2.1.0
+  - @zappjs/json@0.2.0
+
 ## 0.3.0
 
 ### Minor Changes